fix ohiru && reformat files
